Cordelia testing and figuring things out
I copied and wrote down the code that every person's file had, because I am assuming that it is important in making the code function properly. I also copied the subsystems and the commands examples so that I can use them to help me later. Then I read the code hoping to learn about what it is doing. Finally I working on figuring out how to commit code.